facial_recognition-main/requirements_3.10.txt
mediapipe==0.10.11
deepface==0.0.83
opencv-python==4.9.0.80
tensorflow==2.13.0
tf-keras==2.15.0
pyserial
django
gunicorn